refactor seperating stars component from create
중요도 컴포넌트 분리했습니다.
